Don't accept PRs which are in draft status (#6)
This fixes both a UX issue and a correctness issue

The UX improvement is that a draft PR is one that is probably not ready
for inclusion.  It should be moved out of draft status.

The correctness issue is that if squash merging is enabled then the
squash merge via GitHub's API will fail if the PR is still in draft
status.